+++ memory-game/js/modules/delete-template.js
@@ -1,0 +1,6 @@
+export const deleteTableTemplate = () => {
+    const tableRows = document.querySelectorAll('.score__row');
+    if (tableRows) {
+        tableRows.forEach(row => row.remove());
+    }
+};
