+++ memory-game/js/utils/shuffle.js
@@ -1,0 +1,6 @@
+export const shuffle = (array, number) => {
+    array.forEach(elem => {
+        let ramdomPos = Math.floor(Math.random() * number);
+        elem.style.order = ramdomPos;
+    });
+};
